refactor(theme): refactored store switcher
