+++ packages/theme/components/StoreSwitcher.vue
@@ -19,102 +19,50 @@
         </template>
       </SfCharacteristic>
     </SfButton>
-    <SfBottomModal
-      :is-open="isLangModalOpen"
-      :title="availableStores.length > 0 ? 'Change Store': ''"
-      @click:close="isLangModalOpen = !isLangModalOpen"
-    >
-      <SfList
-        v-if="availableStores.length > 1"
-      >
-        <SfListItem
-          v-for="store in availableStores"
-          :key="store.id"
-        >
-          <a
-            href="/"
-            class="container__store--link"
-            :class="storeGetters.getSelected(storeConfig, store) ? 'container__store--selected' : ''"
-            @click="handleChanges({
-              callback: () => changeStore(store),
-              redirect: false,
-              windowRefresh: true,
-            })"
-          >
-            <SfCharacteristic class="language">
-              <template #title>
-                <span>{{ storeConfigGetters.getName(store) }}</span>
-              </template>
-              <template #icon>
-                <nuxt-img
-                  :src="`/icons/langs/${storeConfigGetters.getLocale(store)}.webp`"
-                  width="20"
-                  height="20"
-                  alt="Flag"
-                  class="language__flag"
-                />
-              </template>
-            </SfCharacteristic>
-          </a>
-        </SfListItem>
-      </SfList>
-    </SfBottomModal>
+    <LazyHydrate when-visible>
+      <StoresModal
+        v-if="isLangModalOpen"
+        :store-config="storeConfig"
+        :is-lang-modal-open="isLangModalOpen"
+        @closeModal="isLangModalOpen = false"
+      />
+    </LazyHydrate>
   </div>
 </template>
 
 <script>
+import LazyHydrate from 'vue-lazy-hydration';
 import {
   useConfig,
-  useStore,
   storeConfigGetters,
   storeGetters,
 } from '@vue-storefront/magento';
 import {
   SfButton,
-  SfList,
-  SfBottomModal,
   SfCharacteristic,
 } from '@storefront-ui/vue';
 import {
   ref,
-  computed,
   defineComponent,
-  onMounted,
 } from '@nuxtjs/composition-api';
-import { useHandleChanges } from '~/helpers/magentoConfig/handleChanges';
+import StoresModal from '~/components/StoreSwitcher/StoresModal.vue';
 
 export default defineComponent({
   name: 'StoreSwitcher',
   components: {
+    StoresModal,
     SfButton,
-    SfList,
-    SfBottomModal,
     SfCharacteristic,
+    LazyHydrate,
   },
   setup() {
     const {
       config,
     } = useConfig();
 
-    const {
-      stores,
-      change: changeStore,
-      load: loadStores,
-    } = useStore();
-
-    const { handleChanges } = useHandleChanges();
     const isLangModalOpen = ref(false);
 
-    const availableStores = computed(() => stores.value ?? []);
-
-    onMounted(() => {
-      loadStores();
-    });
-
     return {
-      availableStores,
-      changeStore,
-      handleChanges,
       isLangModalOpen,
       storeConfig: config,
       storeConfigGetters,
@@ -125,84 +73,30 @@ export default defineComponent({
 </script>
 
 <style lang="scss" scoped>
+.language {
+  &__flag {
+    margin-right: var(--spacer-sm);
+  }
+}
 
-.container {
+.container__lang {
+  --button-box-shadow: none;
+  background: none;
+  padding: 0 5px;
   display: flex;
-  flex-wrap: nowrap;
   align-items: center;
-  position: relative;
+  opacity: 0.5;
+  border: none;
 
-  .sf-bottom-modal {
-    z-index: 2;
-    left: 0;
-    @include for-desktop {
-      --bottom-modal-height: 100vh;
-    }
+  &:hover,
+  &--selected {
+    opacity: 1;
   }
+}
 
-  .sf-bottom-modal::v-deep .sf-bottom-modal__close {
-    position: var(--circle-icon-position, absolute);
-    top: var(--spacer-xs);
-    right: var(--spacer-xs);
-  }
-
-  .language {
-    padding: var(--spacer-sm);
-
-    &__flag {
-      margin-right: var(--spacer-sm);
-    }
-  }
-
-  .sf-list {
-    @include for-desktop {
-      display: flex;
-    }
-  }
-
-  &__store {
-    &--selected {
-      font-weight: bold;
-    }
-  }
-
-  &__lang {
-    --button-box-shadow: none;
-    background: none;
-    padding: 0 5px;
-    display: flex;
-    align-items: center;
-    opacity: 0.5;
-    border: none;
-
-    &:hover,
-    &--selected {
-      opacity: 1;
-    }
-
-    &--title {
-      --heading-title-font-weight: var(--font-weight--normal);
-      padding: var(
-              --bottom-modal-title-padding,
-              var(--spacer-sm) var(--spacer-lg)
-      );
-      color: var(--bottom-modal-title-color, var(--c-text));
-      text-align: var(--bottom-modal-title-text-align, center);
-      @include for-mobile {
-        --heading-title-font-size: var(--font-size--xs);
-        --heading-title-font-weight: var(--font-weight--bold);
-      }
-    }
-  }
-
-  .sf-characteristic__text {
-    color: var(--bottom-modal-title-color, var(--c-text));
-  }
-
-  .store-switcher {
-    text-transform: capitalize;
-    color: var(--bottom-modal-title-color, var(--c-link));
-    font-size: var(--font-size--sm);
-  }
+.store-switcher {
+  text-transform: capitalize;
+  color: var(--bottom-modal-title-color, var(--c-link));
+  font-size: var(--font-size--sm);
 }
 </style>
